requirements.txt
Pillow
numpy
selenium
tensorflow
sklearn
matplotlib